completed front-back integration

+++ Domain.Payments/Entities/Withdrawal.cs
@@ -3,12 +3,12 @@ namespace Domain.Payments.Entities;
 public class Withdrawal
 {
     public int Id { get; set; }
-    public int Amount { get; set; }
+    public double Amount { get; set; }
 
     public DateTime DateTime { get; set; }
     public Wallet Wallet { get; set; }
     
-    public Withdrawal(int amount, Wallet wallet)
+    public Withdrawal(double amount, Wallet wallet)
     {
         Amount = amount;
         Wallet = wallet;
